requirements.txt
asyncio
google-auth
google-cloud-bigquery
flask
aiogram
dotenv